pyproject.toml
[project]
name = "signgpt-crew-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.152.0",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
]
